+++ src/test/polymer/polymer2-element-scanner_test.ts
@@ -358,5 +358,4 @@ namespaced name.`,
           },
         ]);
       });
-
 });