+++ CHANGELOG.md
@@ -11,10 +11,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    Unreleased section, uncommenting the header as necessary.
 -->
 
-<!--## Unreleased-->
+## Unreleased
 * Added a `_fork` method to Analyzer.
 * Support for method detection on polymer elements and mixins via the new `methods` property.
 * Support for function analysis (namespaced functions only for now, via the `@memberof` jsdoc tag).
+* Track privacy better on properties and methods.
+  * replaced `private: boolean` with `privacy: 'public' | 'private' | 'protected'`
+  * respects `@public` `@private` and `@protected` in jsdoc
+  * considers one leading underscore to be protected, and two to be private
+  * one trailing underscore is private (closure style)
 
 ## [2.0.0-alpha.28] - 2017-02-24
 * Support for `@memberof` jsdoc tag